pyproject.toml
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "flake8-pytestrail"
version = "0.2.1"
description = "Flake8 plugin to check for missing or wrong TestRail test identifiers"
authors = ["Andrey Semakin <and-semakin@ya.ru>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/and-semakin/flake8-pytestrail"
homepage = "https://github.com/and-semakin/flake8-pytestrail"
keywords = [
  "flake8",
  "flake8-plugin",
  "linting",
  "linter",
  "code quality",
  "test",
  "tests",
  "testrail",
  "pytest"
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Framework :: Flake8",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Quality Assurance",
]

[tool.poetry.plugins."flake8.extension"]
TR = "flake8_pytestrail:PyTestRailChecker"

[tool.poetry.dependencies]
python = "^3.6"
flake8 = ">=3.0.0"
attrs = ">=19.2.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
pytest = "^5.4.1"
mypy = "^0.770"
